requirements.txt
pytz
python-telegram-bot
boto3